dorking_tool/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
pydantic
python-dotenv

# detectors’in bağımlılıkları
pdfplumber
requests
pandas
easyocr
PyMuPDF
opencv-python
python-docx
docx2txt
textract
selenium-wire>=5.0.0,<5.2.0
blinker<1.8.0
beautifulsoup4>=4.11.1
tldextract>=3.3.1

selenium>=4.6.0
selenium-wire>=5.0.0,<5.2.0
webdriver-manager>=3.8.4

reportlab>=3.6.6

openpyxl>=3.0.10
xlrd>=2.0.1

numpy>=1.23.0
matplotlib>=3.6.0

torch>=1.12.0
torchvision>=0.13.0
